refactor: move the types / parsers out into shared place
