+++ public/entry.ts
@@ -1,38 +1,9 @@
-import { z } from "https://esm.sh/zod@3.24.3";
 import { fevt, forEach, fromEvent, fwkr, merge } from "./lib/free.ts";
-import { Patch, patch } from "./patch.ts";
+import { patch } from "./patch.ts";
+import { Event, InnerSize } from "./types.ts";
 
 declare const m: HTMLElement;
 const w = fwkr("./worky.ts", { type: "module" });
-
-const InnerSize = z
-  .object({
-    innerWidth: z.number(),
-    innerHeight: z.number(),
-  })
-  .transform(({ innerWidth: width, innerHeight: height }) => ({
-    width,
-    height,
-  }));
-
-const ErrorEvent = z.object({
-  type: z.literal("error"),
-});
-
-const MessageErrorEvent = z.object({
-  type: z.literal("messageerror"),
-});
-
-const MessageEvent = z.object({
-  type: z.literal("message"),
-  data: Patch,
-});
-
-const Event = z.discriminatedUnion("type", [
-  ErrorEvent,
-  MessageErrorEvent,
-  MessageEvent,
-]);
 
 forEach(
   merge([
@@ -44,7 +15,7 @@ forEach(
     const { success, data: event, error } = Event.safeParse(e);
     if (!success) {
       console.error(error);
-      console.info({ data: e });
+      console.info({ event: e });
       return;
     }
