+++ public/worky.tsx
@@ -2,7 +2,6 @@ import { enablePatches, produceWithPatches } from "https://esm.sh/immer@10.1.1";
 import isEqual from "https://esm.sh/lodash-es@4.17.21/isEqual.js";
 import isPlainObject from "https://esm.sh/lodash-es@4.17.21/isPlainObject.js";
 import mergeWith from "https://esm.sh/lodash-es@4.17.21/mergeWith.js";
-import { z } from "https://esm.sh/zod@3.24.3";
 import { App } from "./components/App.tsx";
 import {
   fcev,
@@ -14,23 +13,9 @@ import {
   type Last,
 } from "./lib/free.ts";
 import type { El } from "./lib/real.ts";
+import { Msg } from "./types.ts";
 
 type Customizer = Last<Parameters<typeof mergeWith<any, any>>>;
-
-const RenderEvent = z.object({
-  type: z.literal("render"),
-});
-
-const ResizeEvent = z.object({
-  type: z.literal("resize"),
-  width: z.number(),
-  height: z.number(),
-});
-
-const Event = z.discriminatedUnion("type", [
-  RenderEvent,
-  ResizeEvent,
-]);
 
 const customizer: Customizer = (a, b, k) => {
   if (k === "props" && isPlainObject(a)) {
@@ -48,7 +33,7 @@ const produce = produceWithPatches((draft, state) => {
   mergeWith(draft, state, customizer);
 });
 
-const evt = ftgt();
+const tgt = ftgt();
 let tree: El = { tag: null };
 
 const render = (nextTreeFn: any) => {
@@ -64,12 +49,12 @@ const render = (nextTreeFn: any) => {
 
   tree = prodTree;
   patches.forEach((patch) =>
-    evt.dispatchEvent(fcev("patch", { detail: patch }))
+    tgt.dispatchEvent(fcev("patch", { detail: patch }))
   );
 };
 
 forEach(
-  fromEvent<CustomEvent>(evt, "patch"),
+  fromEvent<CustomEvent>(tgt, "patch"),
   ({ detail }) => self.postMessage(detail),
 );
 
@@ -77,21 +62,21 @@ let treeFn: any;
 forEach(
   fromEvent<MessageEvent>(self, "message"),
   ({ data }) => {
-    const { success, data: event, error } = Event.safeParse(data);
+    const { success, data: msg, error } = Msg.safeParse(data);
     if (!success) {
       console.error(error);
       console.info({ data });
       return;
     }
 
-    switch (event.type) {
+    switch (msg.type) {
       case "render": {
         treeFn && render(treeFn);
         break;
       }
 
       case "resize": {
-        const { width, height } = event;
+        const { width, height } = msg;
         treeFn = <App width={width} height={height} />;
         render(treeFn);
         break;
